remove loose test file
